requirements.txt
annotated-types==0.7.0
anyio==4.9.0
catboost==1.2.7
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
fastapi==0.115.12
fonttools==4.56.0
gensim==4.3.3
graphviz==0.20.3
h11==0.14.0
idna==3.10
joblib==1.4.2
kiwisolver==1.4.8
matplotlib==3.10.1
narwhals==1.33.0
numpy==1.26.4
packaging==24.2
pandas==2.2.3
pillow==11.1.0
plotly==6.0.1
pydantic==2.11.1
pydantic_core==2.33.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
starlette==0.46.1
threadpoolctl==3.6.0
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
uvicorn==0.34.0
wrapt==1.17.2
psycopg2-binary==2.9.9
kagglehub==0.3.10
seaborn
hvac==2.1.0
